\name{PISA2003lite-package}
\alias{PISA2003lite-package}
\alias{PISA2003lite}
\docType{package}
\title{
Set of datasets from PISA 2003 study
}
\description{
This package aggregates datasets from PISA 2003 study. 
The original dataset is avaliable here: \code{http://pisa2003.acer.edu.au/downloads.php}. 
It was converted into the R package to facilitate access for researchers familiar with R.

The PISA2003lite package containes raw datasets avaliable on the PISA webpage. In the package PISA2003 you will find additional datasets, merged datasets, transformed variables, but be aware because it is a quite large package.

From wikipedia: The Programme for International Student Assessment (PISA) is a worldwide study by the Organisation for Economic Co-operation and Development (OECD) in member and non-member nations of 15-year-old school pupils' scholastic performance on mathematics, science, and reading. It was first performed in 2000 and then repeated every three years. It is done with view to improving education policies and outcomes. The data has increasingly been used both to assess the impact of education quality on incomes and growth and for understanding what causes differences in achievement across nations.
}
\details{
\tabular{ll}{
Package: \tab PISA2003lite\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-06-25\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
The dataset is created by OECD. Visit this page for more details \code{http://www.oecd.org/pisa/}.

Maintainer: Przemyslaw Biecek <Przemyslaw.Biecek@gmail.com>
}
\references{
\code{http://www.oecd.org/pisa/}
}
\keyword{ package }
\seealso{
Other releases of PISA \code{\link{PISA2000lite}}, \code{\link{PISA2006lite}}, \code{\link{PISA2009lite}}, \code{\link{PISA2012lite}}.
}
\examples{
## Do not run
#demo(PISA2003lite)
}
